FIX(plugins): Unset active positional plugin on unloading it
If a plugin is providing positional audio data and it becomes unloaded,
such as by disabling the plugin in the settings GUI, the plugin manager
may still interact with the plugin through m_activePositionalDataPlugin.

This looks unintentional and, in debug mode, this fails an assertion in
the plugin fairly quickly: `assertPluginLoaded(this)`.

Also, the plugin manager shutdown is moved earlier, since quitting
mumble and shutting down the plugin manager while a plugin is linked
causes it the plugin manager to log a message through the global Log
instance and for the Log instance to display a message on the main
window. So bot the Log and the main window must still exist.